Sonar cloud highlighted issue fixed
