+++ src/app/core/utils/data-flattner.spec.ts
@@ -1,0 +1,7 @@
+import { DataFlattner } from './data-flattner';
+
+describe('DataFlattner', () => {
+  it('should create an instance', () => {
+    expect(new DataFlattner()).toBeTruthy();
+  });
+});
