+++ src/app/organization/rural-outlet/edit-retail-outlet.resolver.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import {
-  Router, Resolve,
+   Resolve,
   ActivatedRoute,
   ActivatedRouteSnapshot
 } from '@angular/router';
-import { Observable, of } from 'rxjs';
+import { Observable } from 'rxjs';
 import { OrganizationService } from '../organization.service';
 
 @Injectable({
@@ -12,7 +12,7 @@ import { OrganizationService } from '../organization.service';
 })
 export class EditRetailOutletResolver implements Resolve<object> {
   constructor(private organizationService: OrganizationService,private route:ActivatedRoute) {}
-  resolve(route: ActivatedRouteSnapshot): Observable<any> {    
+  resolve(route: ActivatedRouteSnapshot): Observable<any> {
     const retailOutletId  = route.paramMap.get('id');
     return this.organizationService.getRuralOutletByOutletId(retailOutletId);
   }
