+++ src/app/organization/rural-outlet/retail-outlet.resolver.ts
@@ -1,10 +1,8 @@
 import { Injectable } from '@angular/core';
 import {
-  Router, Resolve,
-  RouterStateSnapshot,
-  ActivatedRouteSnapshot
+   Resolve,
 } from '@angular/router';
-import { Observable, of } from 'rxjs';
+import { Observable } from 'rxjs';
 import { OrganizationService } from '../organization.service';
 
 @Injectable({
