+++ src/app/organization/rural-outlet/rural-outlet.component.ts
@@ -56,7 +56,7 @@ export class RuralOutletComponent implements OnInit {
 
   deactivateOutlet(outletId: any,status:boolean) {
     const outlet=this.outletData.filter(x=>x.id===outletId);
-    if(status==false){
+    if(!status){
     const disableOutletDialogRef = this.dialog.open(DisableDialogComponent, {
       data: { disableContext: `rural outlet : ${outlet[0]?.name}` }
     });
